refactor(libs): convert to TypeScript
