--- libs/constants/index.d.ts
@@ -1,15 +1,0 @@
-export const ACTIVE_LOCALES: Set<string>;
-export const VALID_LOCALES: Map<string, string>;
-export const RETIRED_LOCALES: Map<string, string>;
-export const DEFAULT_LOCALE: string;
-export const LOCALE_ALIASES: Map<string, string>;
-export const PREFERRED_LOCALE_COOKIE_NAME: string;
-export const CSP_SCRIPT_SRC_VALUES: string[];
-export const CSP_VALUE: string;
-export const FLAW_LEVELS: Readonly<Record<string, string>>;
-export const VALID_FLAW_CHECKS: Set<string>;
-export const MDN_PLUS_TITLE: string;
-export const HTML_FILENAME: string;
-export const MARKDOWN_FILENAME: string;
-export const VALID_MIME_TYPES: Set<string>;
-export const MAX_COMPRESSION_DIFFERENCE_PERCENTAGE: number;
