--- libs/fundamental-redirects/index.d.ts
@@ -1,4 +1,0 @@
-export function resolveFundamental(path: string): {
-  url?: string;
-  status?: 301 | 302;
-};
