--- libs/languages/index.d.ts
@@ -1,3 +1,0 @@
-declare const _exports: Record<string, { English: string; native: string }>;
-
-export default _exports;
