--- libs/locale-utils/index.d.ts
@@ -1,2 +1,0 @@
-export function getLocale(request: any, fallback?: string): any;
-export function isValidLocale(locale: any): locale is string;
