+++ libs/locale-utils/package.json
@@ -4,9 +4,8 @@
   "private": true,
   "license": "MPL-2.0",
   "type": "module",
-  "exports": "./index.js",
-  "main": "index.js",
-  "types": "index.d.ts",
+  "exports": "./index.ds",
+  "main": "index.ds",
   "dependencies": {
     "accept-language-parser": "^1.5.0",
     "cookie": "^0.5.0"
