--- libs/pong/cc2ip.d.ts
@@ -1,1 +1,0 @@
-export default function anonymousIpByCC(countryCode: string): string;
