--- libs/pong/click.d.ts
@@ -1,8 +1,0 @@
-import { Coder } from "./coding.js";
-
-export function createPongClickHandler(coder: Coder): (
-  params: URLSearchParams
-) => Promise<{
-  status: number;
-  location: string;
-}>;
