--- libs/pong/image.d.ts
@@ -1,4 +1,0 @@
-export function fetchImage(src: string): Promise<{
-  buf: ArrayBuffer;
-  contentType: string;
-}>;
