+++ libs/pong/image.ts
@@ -1,5 +1,7 @@
 /* global fetch */
-export async function fetchImage(src) {
+export async function fetchImage(
+  src: string
+): Promise<{ buf: ArrayBuffer; contentType: string }> {
   const imageResponse = await fetch(src);
   const imageBuffer = await imageResponse.arrayBuffer();
   const contentType = imageResponse.headers.get("content-type");
