--- libs/pong/index.js
@@ -1,6 +1,0 @@
-export * from "./coding.js";
-export * from "./get.js";
-export * from "./image.js";
-export * from "./click.js";
-export * from "./viewed.js";
-export * from "./fallback.js";
