--- libs/pong/viewed.d.ts
@@ -1,5 +1,0 @@
-import { Coder } from "./coding.js";
-
-export function createPongViewedHandler(
-  coder: Coder
-): (params: URLSearchParams) => Promise<void>;
