+++ libs/pong/viewed.ts
@@ -1,5 +1,9 @@
+import { Coder } from "./coding.js";
+
 /* global fetch */
-export function createPongViewedHandler(coder) {
+export function createPongViewedHandler(
+  coder: Coder
+): (params: URLSearchParams) => Promise<void> {
   return async (params) => {
     const view = coder.decodeAndVerify(params.get("code"));
     const fallback = coder.decodeAndVerify(params.get("fallback"));
