--- libs/slug-utils/index.d.ts
@@ -1,3 +1,0 @@
-export function slugToFolder(slug: string, joiner?: string): string;
-export function decodePath(path: string): string;
-export function encodePath(path: string): string;
